sim: Add device with larger sectors in slot 0 than in slot 1
The simulator was not performing any test with a configuration where the
primary slot is composed of larger sectors than the secondary slot.
This can be typically case when using a STM32 with an external flash
memory since most STM32 have large internal sectors. This configuration
was causing issues when using the swap-scratch upgrade strategy.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>